Revert "update repo"

+++ plugins/Pulsars/src/Pulsar.hpp
@@ -104,7 +104,10 @@ public:
 	//! Get the english name of pulsar
 	virtual QString getEnglishName(void) const;
 	//! Get the designation of pulsar
-	QString getDesignation(void) const { return designation; }
+	QString getDesignation(void) const
+	{
+		return designation;
+	}
 
 	void update(double deltaTime);
 
@@ -127,8 +130,8 @@ private:
 	//! Variables for description of properties of pulsars
 	QString designation;	//! The designation of the pulsar (J2000 pulsar name)
 	QString pulsarName;	//! The proper name of the pulsar
-	double RA;		//! J2000 right ascension
-	double DE;		//! J2000 declination
+	float RA;		//! J2000 right ascension
+	float DE;		//! J2000 declination
 	float parallax;		//! Annual parallax (mas)
 	double period;		//! Barycentric period of the pulsar (s)
 	double frequency;	//! Barycentric rotation frequency (Hz)
@@ -160,6 +163,7 @@ private:
 	//! Get type of pulsar
 	//! @param pcode - code of pulsar type
 	QString getPulsarTypeInfoString(QString pcode) const;
+
 };
 
 #endif // PULSAR_HPP
